refactor simulation init system
