+++ app/index.html
@@ -16,63 +16,66 @@
         href="https://github.com/ubernaut/spaceSim"
         target="_blank"
         rel="noopener"
-        >On GitHub</a
-      >
+        >On GitHub
+      </a>
     </div>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+
     <div id="help">
       <h3 onclick="$('#controls').toggle();">(click here to toggle help)</h3>
       <pre id="controls">
-    <h3>General Info:</h3>
-    This is a nbody Gravity simulation using GPU.js and
-    THREE.js. The physics is being computed on the GPU
-    and there is still a lot of room for improvment.
+        <h3>General Info:</h3>
+        This is a nbody Gravity simulation using GPU.js and
+        THREE.js. The physics is being computed on the GPU
+        and there is still a lot of room for improvment.
 
-    To change the number of bodies in the simulation,
-    change the "bodyCount" url parameter thusly:
+        To change the number of bodies in the simulation,
+        change the "bodyCount" url parameter thusly:
 
-    <a class="brightText" href="http://thedagda.co:9000?bodyCount=1000">http://thedagda.co:9000?bodyCount=1000</a>
-    you can also disable stars by setting "nostars=true":
-    <a class="brightText" href="http://thedagda.co:9000?bodyCount=100&stars=false">http://thedagda.co:9000?bodyCount=100&stars=false</a>
-    <h3>Controls:</h3>
-    wasd  (as is tradition)
-    q,e to roll
-    r,f for up and down
+        <a class="brightText" href="http://thedagda.co:9000?bodyCount=1000">http://thedagda.co:9000?bodyCount=1000</a>
+        you can also disable stars by setting "nostars=true":
+        <a class="brightText" href="http://thedagda.co:9000?bodyCount=100&stars=false">http://thedagda.co:9000?bodyCount=100&stars=false</a>
+        <h3>Controls:</h3>
+        wasd  (as is tradition)
+        q,e to roll
+        r,f for up and down
 
-    Click and drag the mouse to rotate your view.
+        Click and drag the mouse to rotate your view.
 
-    spacebar shoots stuff (try holding it down)
+        spacebar shoots stuff (try holding it down)
 
-    Mousewheel zooms
-    --------------------
-    You can zoom out very far with the
-    logarithmic Depth Buffer.
+        Mousewheel zooms
+        --------------------
+        You can zoom out very far with the
+        logarithmic Depth Buffer.
 
-    your ship is 20 meters across & the planets
-    and star are to scale
+        your ship is 20 meters across & the planets
+        and star are to scale
 
-    if you zoom way out you will see a number of large spheres.
+        if you zoom way out you will see a number of large spheres.
 
-    The Gray sphere is the size of
-    our solar system (oort cloud)
+        The Gray sphere is the size of
+        our solar system (oort cloud)
 
-    Past that you'll see the stars moving
-    they're spread out around an area
-    roughly equal the to milky way.
+        Past that you'll see the stars moving
+        they're spread out around an area
+        roughly equal the to milky way.
 
-    The white sphere is the size of
-    the milky way.
+        The white sphere is the size of
+        the milky way.
 
-    The red sphere is the size of the
-    observable universe.
-    ----------------------
-    This is multiplayer
-    you may see other ships,
-    you may shoot them with the colored balls
-    ----------------------
-    </pre>
+        The red sphere is the size of the
+        observable universe.
+        ----------------------
+        This is multiplayer
+        you may see other ships,
+        you may shoot them with the colored balls
+        ----------------------
+      </pre>
     </div>
+
     <div id="root"></div>
+
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
     <script src="app/lib/gpu.js"></script>
     <script src="/app/lib/three/three.js"></script>
     <script src="/app/lib/three/loaders/DDSLoader.js"></script>
