+++ app/js/sim/index.js
@@ -1,0 +1,3 @@
+import init from './init'
+
+export default init
