+++ app/js/sim/physics.js
@@ -1,0 +1,27 @@
+const updateOimoPhysics = () => {
+  if (Void.world == null) {
+    return
+  }
+  Void.world.step()
+}
+
+const initOimoPhysics = () => {
+  // world setting:( TimeStep, BroadPhaseType, Iterations )
+  // BroadPhaseType can be
+  // 1 : BruteForce
+  // 2 : Sweep and prune , the default
+  // 3 : dynamic bounding volume tree
+  return new OIMO.World({
+    timestep: 1 / 60,
+    iterations: 8,
+    broadphase: 2,
+    worldscale: 1,
+    random: true,
+    info: false,
+    gravity: [ 0, 0, 0 ]
+  })
+  // populate(1);
+  // setInterval(updateOimoPhysics, 1000/60);
+}
+
+export { updateOimoPhysics, initOimoPhysics }
