+++ app/js/sim/system.js
@@ -1,28 +1,12 @@
 import Promise from 'bluebird'
 
-import { createStar, createRandomStar } from '-/bodies/star'
-import { createPlanet } from '-/bodies/planet'
-import { createUniverse } from '-/bodies/universe'
-import { createGalaxy, addStars } from '-/bodies/galaxy'
-import * as controls from '-/player/controls'
-import { createGamepadControls } from '-/player/controls/gamepad-controls'
-import * as weapons from '-/player/weapons'
-import { createShip } from '-/player/ship'
-import { createDrone } from '-/player/drone'
-import { getAllConfigVars, randomUniform } from '-/utils'
-
 import soPhysics from '@void/core/system-builder/soPhysics'
 import { convertSystemToMeters } from '@void/core/system-builder/utils'
 
+import { createRandomStar } from '-/bodies/star'
+import { createPlanet } from '-/bodies/planet'
+import { randomUniform } from '-/utils'
 import SystemBuilderWorker from './workers/systemBuilder.worker'
-// import {getConfig} from './systemBuilder'
-// import { Clock, PerspectiveCamera, Scene, WebGLRenderer } from 'three'
-import { EffectComposer, BloomPass, RenderPass } from 'postprocessing'
-
-let renderer
-// let world = null
-
-const animateCallbacks = []
 
 const loadSystem = () => {
   let bodyCount = 1024
@@ -159,7 +143,7 @@ const loadSystem = () => {
         // Void.scene.add(star.photosphere)
         Void.scene.add(star.chromosphere)
         // Void.scene.add()
-        animateCallbacks.push(star.animate)
+        Void.animateCallbacks.push(star.animate)
       } else {
         const planet = createPlanet({
           radius: body.radius,
@@ -269,208 +253,4 @@ const updateSystemGPU = () => {
   }
 }
 
-const initOimoPhysics = () => {
-  // world setting:( TimeStep, BroadPhaseType, Iterations )
-  // BroadPhaseType can be
-  // 1 : BruteForce
-  // 2 : Sweep and prune , the default
-  // 3 : dynamic bounding volume tree
-  return new OIMO.World({
-    timestep: 1 / 60,
-    iterations: 8,
-    broadphase: 2,
-    worldscale: 1,
-    random: true,
-    info: false,
-    gravity: [ 0, 0, 0 ]
-  })
-  // populate(1);
-  // setInterval(updateOimoPhysics, 1000/60);
-}
-
-const addLights = scene => {
-  const ambient = new THREE.AmbientLight(0x888888)
-  scene.add(ambient)
-}
-
-const squareGrid = () => {
-  const size = 100000000
-  const divisions = 1000
-  const gridHelper1 = new THREE.GridHelper(size, divisions, 0xffffff, 0xfffff)
-}
-
-const addPostprocessing = ({ renderer, scene, camera }) => {
-  // postprocessing
-  composer = new EffectComposer(renderer)
-  composer.addPass(new RenderPass(scene, camera))
-
-  const bloomPass = new BloomPass({
-    resolutionScale: 0.05,
-    kernelSize: 3.0,
-    intensity: 0.3,
-    distinction: 1
-  })
-  bloomPass.renderToScreen = true
-  bloomPass.combineMaterial.defines.SCREEN_MODE = '1'
-  bloomPass.combineMaterial.needsUpdate = true
-  composer.addPass(bloomPass)
-}
-
-const addUniverse = scene => {
-  const oortGeometry = new THREE.SphereGeometry(7.5 * Math.pow(10, 15), 32, 32)
-  const oortMaterial = new THREE.MeshBasicMaterial({ color: 0x555555 })
-  const oort = new THREE.Mesh(oortGeometry, oortMaterial)
-  // scene.add(oort)
-
-  const galaxyGeometry = new THREE.SphereGeometry(5 * Math.pow(10, 20), 32, 32)
-  const galaxyMaterial = new THREE.MeshBasicMaterial({ color: 0xffffff })
-  const galaxy = new THREE.Mesh(galaxyGeometry, galaxyMaterial)
-  // scene.add(galaxy)
-
-  const universeGeometry = new THREE.SphereGeometry(
-    4.4 * Math.pow(10, 26),
-    32,
-    32
-  )
-  const universeMaterial = new THREE.MeshBasicMaterial({ color: 0xff0000 })
-  const universe = new THREE.Mesh(universeGeometry, universeMaterial)
-  // scene.add(universe)
-
-  return [ oort, galaxy, universe ]
-}
-
-const deployDrone = ship => createDroneOpts => {
-  const drone = createDrone(createDroneOpts)
-  ship.add(drone.mesh)
-  drone.mesh.position.set(5, 5, 5)
-}
-
-let composer
-const init = rootEl => {
-  // renderer
-  renderer = new THREE.WebGLRenderer({
-    antialias: true,
-    logarithmicDepthBuffer: true
-    // shadowMapEnabled: true
-  })
-  renderer.setPixelRatio(window.devicePixelRatio)
-  renderer.setSize(window.innerWidth, window.innerHeight)
-
-  // camera
-  const IAU = 9.4607 * Math.pow(10, 15)
-  const farClip = 5 * IAU
-  const camera = (Void.camera = new THREE.PerspectiveCamera(
-    65,
-    window.innerWidth / window.innerHeight,
-    0.1,
-    farClip
-  ))
-
-  // scene
-  const scene = (Void.scene = new THREE.Scene())
-  addLights(scene)
-
-  // ship
-  createShip({ controls }).then(({ ship, animate }) => {
-    // console.log(ship)
-    Void.ship = ship
-    Void.ship.add(Void.camera)
-    Void.camera.position.set(0, 10, 30)
-    scene.add(ship)
-    animateCallbacks.push(animate)
-
-    if (Void.urlConfigs.hasOwnProperty('gamepad')) {
-      Void.controls = createGamepadControls(
-        Void.ship,
-        rootEl,
-        weapons.shoot,
-        deployDrone(ship)
-      )
-    } else {
-      Void.controls = controls.setFlyControls({
-        camera: Void.camera,
-        ship: Void.ship,
-        el: rootEl
-      })
-    }
-  })
-
-  // all other matter
-  createUniverse(scene).map(body => scene.add(body))
-
-  // postprocessing
-  addPostprocessing({ renderer, scene, camera })
-
-  // attach to the dom
-  rootEl.appendChild(renderer.domElement)
-  let starflag = true
-  if (Void.urlConfigs.hasOwnProperty('stars')) {
-    if (Void.urlConfigs.stars === 'false') {
-      starflag = false
-    }
-  }
-  if (starflag) {
-    Void.galaxy = createGalaxy()
-    // addStars()
-  }
-
-  window.addEventListener('resize', onWindowResize, false)
-
-  // init physics
-  Void.world = initOimoPhysics()
-
-  loadSystem()
-}
-
-const onWindowResize = () => {
-  Void.camera.aspect = window.innerWidth / window.innerHeight
-  Void.camera.updateProjectionMatrix()
-  renderer.setSize(window.innerWidth, window.innerHeight)
-}
-
-const updateOimoPhysics = () => {
-  if (Void.world == null) {
-    return
-  }
-  Void.world.step()
-}
-
-const animate = () => {
-  requestAnimationFrame(animate)
-
-  const delta = Void.clock.getDelta()
-
-  if (Void.controls) {
-    Void.controls.update(delta)
-  }
-
-  Void.time.value = Void.clock.getElapsedTime()
-
-  weapons.animate(delta, Void.time.value)
-
-  if (Void.soPhysics && Void.systemLoaded) {
-    if (Void.urlConfigs.hasOwnProperty('CPU')) {
-      Void.soPhysics.accelerateCuda()
-      updateSystemCPU()
-    } else {
-      Void.soPhysics.GPUAccelerate()
-      let GPUcollisions = true
-      if (Void.urlConfigs.hasOwnProperty('GPUcollisions')) {
-        GPUcollisions = Void.urlConfigs.GPUcollisions
-      }
-      if (GPUcollisions == true) {
-        updateSystemCPU()
-      } else {
-        updateSystemGPU()
-      }
-      //
-    }
-    // updateOimoPhysics()
-
-    animateCallbacks.map(x => x(delta, Void.time.value))
-  }
-
-  composer.render(delta)
-}
-
-export { init, animate, loadSystem }
+export { updateSystemCPU, updateSystemGPU, loadSystem }
